aide-api/pyproject.toml
[tool.poetry]
name = "aide-api"
version = "0.1.0"
description = "REST API server for AIDE Platform - Serve processed news, policies, and ratings"
authors = ["AIDE Team <team@aide-platform.com>"]
readme = "README.md"
packages = [{include = "aide_api"}]

[tool.poetry.dependencies]
python = "^3.11"
aide-data-core = {path = "../aide-data-core", develop = true}
fastapi = "^0.104.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
httpx = "^0.25.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=aide_api"
